+++ src/transport/irc/irc_ssl_client.erl
@@ -1,14 +1,14 @@
 %%%----------------------------------------------------------------------
-%%% File    : irc_lib_client.erl
+%%% File    : transport/irc/irc_ssl_client.erl
 %%% Author  : 0xAX <anotherworldofworld@gmail.com>
-%%% Purpose : Irc transport client.
+%%% Purpose : Irc client with ssl supporting.
 %%%----------------------------------------------------------------------
--module(irc_lib_client).
+-module(irc_ssl_client).
 
 -behaviour(gen_server).
- 
+
 -export([start_link/5]).
- 
+
 %% gen_server callbacks
 -export([init/1,
          handle_call/3,
@@ -16,7 +16,7 @@
          handle_info/2,
          terminate/2,
          code_change/3]).
- 
+
 % irc client state
 -record(state, {
     % irc nick
@@ -35,31 +35,32 @@
 
 start_link(CallbackModule, Host, Port, Channel, Nick) ->
     gen_server:start_link(?MODULE, [CallbackModule, Host, Port, Channel, Nick], []).
- 
+
 init([CallbackModule, Host, Port, Channel, Nick]) ->
     % try to connect
     gen_server:cast(self(), {connect, Host, Port}),
     % init process internal state
     {ok, #state{login = Nick, host = Host, irc_channel = Channel, callback = CallbackModule}}.
- 
+
 handle_call(_Request, _From, State) ->
     {reply, ignored, State}.
 
 %% @doc Try to connect to irc server and join to channel
 handle_cast({connect, Host, Port}, State) ->
     % Try to connect to irc server
-    case gen_tcp:connect(binary_to_list(Host), Port, [{delay_send, false}, {nodelay, true}]) of
+    case ssl:connect(binary_to_list(Host), Port, [{delay_send, false}, {nodelay, true}]) of
         {ok, Socket} ->
-            gen_tcp:send(Socket, "NICK " ++ binary_to_list(State#state.login) ++ "\r\n"),
+            ssl:send(Socket, "NICK " ++ binary_to_list(State#state.login) ++ "\r\n"),
             % Send user data
-            gen_tcp:send(Socket, "USER " ++ binary_to_list(State#state.login) ++ " some fake info\r\n"),
+            ssl:send(Socket, "USER " ++ binary_to_list(State#state.login) ++ " some fake info\r\n"),
             % Join to channel
-            gen_tcp:send(Socket, "JOIN " ++ binary_to_list(State#state.irc_channel) ++ "\r\n"),
+            ssl:send(Socket, "JOIN " ++ binary_to_list(State#state.irc_channel) ++ "\r\n"),
             % return
             {noreply, State#state{socket = Socket, is_auth = true}};
         {error, Reason} ->
             % Some log
             io:format("ERROR: ~p~n", [Reason]),
+            % return
             {noreply, State}
         end;
 
@@ -70,7 +71,7 @@ handle_cast({send_message, Message}, State) ->
     % Send messages
     lists:foreach(fun(Mes) -> 
                       % Send message to irc
-                      gen_tcp:send(State#state.socket, "PRIVMSG " ++ binary_to_list(State#state.irc_channel) ++ " :" ++ Mes ++ "\r\n")
+                      ssl:send(State#state.socket, "PRIVMSG " ++ binary_to_list(State#state.irc_channel) ++ " :" ++ Mes ++ "\r\n")
                   end, 
                   MessagesList),
     % return
@@ -80,12 +81,12 @@ handle_cast(_Msg, State) ->
     {noreply, State}.
 
 %% @doc Incoming message
-handle_info({tcp, Socket, Data}, State) ->
+handle_info({ssl, Socket, Data}, State) ->
     % Parse incoming data
     case string:tokens(Data, " ") of
         ["PING" | _] ->
             % Send pong
-            gen_tcp:send(Socket, "PONG :" ++ binary_to_list(State#state.host) ++ "\r\n");
+            ssl:send(Socket, "PONG :" ++ binary_to_list(State#state.host) ++ "\r\n");
         [_User, "PRIVMSG", _Channel | Message] ->
             % Get incoming message
             [_ | IncomingMessage] = string:join(Message, " "),
@@ -97,6 +98,7 @@ handle_info({tcp, Socket, Data}, State) ->
     % return
     {noreply, State#state{socket = Socket}};
 
+
 handle_info({tcp_closed, _}, State) ->
     % stop and return state
     {stop, normal, State};
@@ -106,7 +108,8 @@ handle_info({tcp_error, _Socket, Reason}, State) ->
     % stop and return state
     {stop, normal, State};
 
-handle_info(_Info, State) ->
+handle_info(Info, State) ->
+    io:format("Info ~p~n", [Info]),
     {noreply, State}.
  
 terminate(_Reason, State) ->